+++ packages/terra-date-picker/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Removed
+* Removed time-picker code from react-datepicker.
 
 4.12.0 - (September 6, 2019)
 ------------------
@@ -66,7 +68,7 @@ Unreleased
 * Minor dependency version bump
 
 4.5.0 - (July 2, 2019)
-------------------      
+------------------
 ### Added
 * Allow terra-date-picker to be a controlled component.
 * Added functionality to announce the pre-selected date to screen readers when navigating within the date picker.